Melhora na interface

+++ soma.c
@@ -2,8 +2,8 @@
 #include<string.h>
 
 int main () {
-	int n1, n2;
-	scanf("%d %d", &n1, &n2);
-	printf("%d\n", n1+n2);
+	long long int n1, n2;
+	scanf("%lld %lld", &n1, &n2);
+	printf("%lld\n", n1+n2);
 	return 0;
 }
